FIX: Facebook Provider give a bigger number, ADD: Redirect to profile page on facebook association, only show social media which isnt yet associated

+++ database/migrations/2020_03_21_183224_create_socials_table.php
@@ -17,7 +17,7 @@ class CreateSocialsTable extends Migration
             $table->id();
             $table->integer('artist_id')->unsigned();
             $table->enum('provider', ['facebook', 'instagram', 'youtube', 'twitter', 'linkedin', 'reddit', 'github']);
-            $table->integer('provider_id')->unsigned();
+            $table->bigInteger('provider_id')->unsigned();
             $table->text('data')->nullable();
             $table->timestamps();
         });
